Harden src/mux compiler-warning posture across the codebase
Fixes every site flagged by a broad set of stricter Clang warnings that
are about to be enabled for netmux: -Wswitch-enum, -Wunused-macros,
-Wfloat-equal, -Wpedantic (retyping void* dispatch-table handlers to a
GenericFnPtr union instead of relying on non-portable object/function
pointer casts), -Wcast-align, -Wdouble-promotion, -Wcast-qual/
-Wwrite-strings (constifying string-literal-only fields and params:
NameTable.name, CMDENT.cmdname, CONF.pname, log_error's message args,
etc.), and -Wconversion/-Wsign-conversion (mistyped bitmask constants
that used the sign bit, e.g. SW_MULTIPLE/CF_DARK/LOG_ALWAYS now stored
as int instead of unsigned int; assorted narrowing casts on bounded
DbRef/size_t/time_t values).

Also applies the two required shared-header fixes in src/btech
(wildcard_match's declaration, glue.c's cmdsave local) needed to keep
btech compiling against the now-const mudstate.debug_cmd and
quick_wild signature; btech itself is not yet held to any of these
new warning flags.

+++ src/mux/database/vattr.c
@@ -196,7 +196,7 @@ static char *store_string(char *str) {
   int len;
   char *ret;
 
-  len = strlen(str);
+  len = (int)strlen(str);
 
   /*
    * If we have no block, or there's not enough room left in the * * *
